+++ IngestrTests/IngestrTests.swift
@@ -1,0 +1,16 @@
+//
+//  IngestrTests.swift
+//  IngestrTests
+//
+//  Created by David Klee on 5/18/25.
+//
+
+import Testing
+
+struct IngestrTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}